+++ NEWS.md
@@ -1,3 +1,5 @@
+# fs (development version)
+
 # fs 1.3.0
 
 ## Breaking changes