+++ NEWS.md
@@ -3,6 +3,8 @@
 * fs now passes along CPPFLAGS during compilation of libuv, fixing an issue that could
   prevent compilation from source on macOS Catalina. (@kevinushey, #229)
   
+* `path()` now errors if the path given or constructed will exceed `PATH_MAX` (#233).
+
 * `dir_tree()` now works with paths that need tilde expansion (@dmurdoch, @jennybc, #203).
 * `is_dir()`, `is_file()`, `is_file_empty()` and `file_info()` gain a `follow`
   argument, to follow links and return information about the linked file rather