+++ NEWS.md
@@ -1,4 +1,4 @@
-# fs (development version)
+# fs 1.3.2
 
 * fs now passes along `CPPFLAGS` during compilation of libuv, fixing an issue that could
   prevent compilation from source on macOS Catalina. (@kevinushey, #229)